modules/gcp_infrastructure/python_scripts/function_code/mysql/requirements.txt
async-timeout==4.0.2
backoff==2.2.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
Deprecated==1.2.13
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
mysql-connector==2.2.9
opentelemetry-api==1.16.0
opentelemetry-distro==0.37b0
opentelemetry-exporter-otlp==1.16.0
opentelemetry-exporter-otlp-proto-grpc==1.16.0
opentelemetry-exporter-otlp-proto-http==1.16.0
opentelemetry-instrumentation==0.37b0
opentelemetry-instrumentation-aws-lambda==0.37b0
opentelemetry-instrumentation-dbapi==0.37b0
opentelemetry-instrumentation-grpc==0.37b0
opentelemetry-instrumentation-logging==0.37b0
opentelemetry-instrumentation-mysql==0.37b0
opentelemetry-instrumentation-redis==0.37b0
opentelemetry-instrumentation-requests==0.37b0
opentelemetry-instrumentation-sqlite3==0.37b0
opentelemetry-instrumentation-urllib==0.37b0
opentelemetry-instrumentation-urllib3==0.37b0
opentelemetry-instrumentation-wsgi==0.37b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.16.0
opentelemetry-sdk==1.16.0
opentelemetry-semantic-conventions==0.37b0
opentelemetry-util-http==0.37b0
packaging==23.0
proto-plus==1.22.2
protobuf==4.22.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
redis==4.4.2
requests==2.28.2
rsa==4.9
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.14
wrapt==1.14.1
